requirements.txt
argcomplete==1.9.3
argh==0.26.2
awsebcli==3.17.0
blessed==1.17.2
boto3==1.10.45
botocore==1.13.45
cached-property==1.5.1
cement==2.8.2
certifi==2019.11.28
cfn-flip==1.2.2
chardet==3.0.4
Click==7.0
colorama==0.3.9
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.1.1
django-cors-headers==3.2.1
django-crontab==0.7.1
django-import-export==2.0
django-livesync==0.4
django-restframework==0.0.1
django-s3-storage==0.12.4
django-sslserver==0.22
django-storages==1.9.1
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
docker==3.7.3
docker-compose==1.23.2
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.15.2
drf-yasg==1.17.0
durationpy==0.5
et-xmlfile==1.0.1
future==0.16.0
hjson==3.0.1
idna==2.7
inflection==0.3.1
itypes==1.1.0
jdcal==1.4.1
Jinja2==2.10.3
jmespath==0.9.3
jsonschema==2.6.0
kappa==0.6.0
lambda-packages==0.20.0
MarkupPy==1.14
MarkupSafe==1.1.1
odfpy==1.4.0
openpyxl==3.0.2
packaging==19.2
pathspec==0.5.9
pathtools==0.1.2
placebo==0.9.0
PyJWT==1.7.1
pymysql
pynamodb==4.2.0
pyparsing==2.4.6
python-dateutil==2.6.1
python-slugify==1.2.4
pytz==2019.3
PyYAML==3.13
requests==2.20.1
ruamel.yaml==0.16.5
ruamel.yaml.clib==0.2.0
s3transfer==0.2.1
semantic-version==2.5.0
six==1.11.0
tablib==0.14.0
termcolor==1.1.0
texttable==0.9.1
toml==0.10.0
tornado==4.5.1
tqdm==4.19.1
travis==0.0.3
troposphere==2.5.3
Unidecode==1.1.1
uritemplate==3.0.1
urllib3==1.24.3
watchdog==0.9.0
wcwidth==0.1.8
websocket-client==0.57.0
Werkzeug==0.16.0
wsgi-request-logger==0.4.6
xlrd==1.2.0
xlwt==1.3.0
zappa==0.48.2
